Update v1.1 documentation
The chamber thermistor is now a key recommendation for MissChanger, as a mean to monitor and enable functions that control the chamber temp.
Nevertheless, the means to control chamber temp is (i.e. Nevermore StealthMax) is still in development, and will remained as "optional" even after it is done. This is because it is a completely different project to MissChanger, with different dev and print parameters.