Update toolchanger-calibrate.cfg
Separate the calibration of gcode_offsets (CALIBRATE_RGO) and absolute z_offset (CALIBRATE_AZO).
This is due to the flaky nature of the Nudge probe, when it is built with stainless steel bolts. The long stroke during AZO can cause the probe to misaligned and fail. To mitigate this problem from the software side, the two funtions has been seperated. Furthermore, the AZO has been simplified, to require manual trigger for each toolhead It is recommended to manualy reset the probe (by hand) to between AZO.